requirements.txt
click>=8.1.0,<8.2.0
requests>=2.0.0,<3.0.0
PyYAML
graphql-core>=3.1.0